Form: rearrange form tests
