+++ packages/devextreme/testing/tests/DevExpress.ui.widgets.form/form.API.registerKeyHandler.tests.js
@@ -1,6 +1,10 @@
 import $ from 'jquery';
+import device from 'core/devices';
 
 import 'ui/form';
+import registerKeyHandlerTestHelper from '../../helpers/registerKeyHandlerTestHelper.js';
+
+const EDITOR_INPUT_CLASS = 'dx-texteditor-input';
 
 QUnit.testStart(function() {
     const markup = '<div id="form"></div>';
@@ -32,3 +36,19 @@ QUnit.test('Set { items: [{dataField}] }, call registerKeyHandler', function(ass
     form.registerKeyHandler('tab', handler);
     assert.ok(true, 'no exceptions');
 });
+
+if(device.current().deviceType === 'desktop') {
+    const items = [
+        { dataField: 'name', editorType: 'dxTextBox' },
+        { dataField: 'age', editorType: 'dxNumberBox' }
+    ];
+
+    items.forEach((item) => {
+        registerKeyHandlerTestHelper.runTests({
+            createWidget: ($element) => $element.dxForm({ items: items }).dxForm('instance'),
+            keyPressTargetElement: (widget) => widget.getEditor(item.dataField).$element().find(`.${EDITOR_INPUT_CLASS}`),
+            checkInitialize: false,
+            testNamePrefix: `Form -> ${item.editorType}:`
+        });
+    });
+}
